mini-os: initial (debugging) version of modified netfront for nf10.
It is a working version, but ineffective and unoptimized, for nfback,
an integrated netback in nf10 lbuf nic. APIs for lbuf is provided by
include/nf10_lbuf_api.h. One of unoptimized features is for a packet
that spans across two possibly non-contiguous pages should be copied to
a temporal buffer, which is then passed to upper layer. It will be
optimized by using contigous pages preallocated for consecutive RX ring
entries. Also printing debugging messages will be taken out, since it is
the main source of low performance, but it helps to debug backend's
bookkeeping function for in-flight requests pending for unavailability
of RX rings, which are way limited.